Tests in separated files, regex in game_files
What's going on:
- new undocumented way to parse game files
- parse(abilities) causes end of recursion stack
- tests in separated files
