+++ atod/tests/tests_data/game_files/intends.json
@@ -1,12 +1,8 @@
 {
     "item_dragon_lance": {
-        "ItemQuality": "artifact",
         "ItemCost": "1900",
+        "ID": "236",
         "AbilitySpecial": {
-            "03": {
-                "base_attack_range": "140",
-                "var_type": "FIELD_INTEGER"
-            },
             "01": {
                 "bonus_agility": "13",
                 "var_type": "FIELD_INTEGER"
@@ -14,12 +10,16 @@
             "02": {
                 "bonus_strength": "12",
                 "var_type": "FIELD_INTEGER"
+            },
+            "03": {
+                "base_attack_range": "140",
+                "var_type": "FIELD_INTEGER"
             }
         },
-        "ItemShopTags": "damage",
+        "ItemAliases": "dragon lance",
+        "ItemQuality": "artifact",
         "ItemDisassembleRule": "DOTA_ITEM_DISASSEMBLE_ALWAYS",
-        "AbilityBehavior": "DOTA_ABILITY_BEHAVIOR_PASSIVE",
-        "ItemAliases": "dragonlance",
-        "ID": "236"
+        "ItemShopTags": "damage",
+        "AbilityBehavior": "DOTA_ABILITY_BEHAVIOR_PASSIVE"
     }
 }
